Make compiler.js bundle runnable
- Extracts buildDoms to their own file to reduce import-based side-effects.
- Creates a `self` polyfill
