+++ src/builtins/amp-layout/amp-layout.js
@@ -1,10 +1,10 @@
-import {Layout, applyFillContent, isLayoutSizeDefined} from '#core/dom/layout';
-import {realChildNodes} from '#core/dom/query';
+import {Layout, isLayoutSizeDefined} from '#core/dom/layout';
 
 import {registerElement} from '#service/custom-element-registry';
 
+import {buildDom} from './build-dom';
+
 import {BaseElement} from '../../base-element';
-import {getEffectiveLayout} from '../../static-layout';
 
 export class AmpLayout extends BaseElement {
   /** @override @nocollapse */
@@ -21,26 +21,6 @@ export class AmpLayout extends BaseElement {
   buildCallback() {
     buildDom(this.element);
   }
-}
-
-/**
- * @see amphtml/compiler/types.js for full description
- *
- * @param {!Element} element
- */
-export function buildDom(element) {
-  const layout = getEffectiveLayout(element);
-  if (layout == Layout.CONTAINER) {
-    return;
-  }
-
-  const doc = element.ownerDocument;
-  const container = doc.createElement('div');
-  applyFillContent(container);
-  realChildNodes(element).forEach((child) => {
-    container.appendChild(child);
-  });
-  element.appendChild(container);
 }
 
 /**
