+++ src/compiler/builders.js
@@ -1,6 +1,6 @@
-import {buildDom as ampLayoutClassic} from '#builtins/amp-layout/amp-layout';
+import {buildDom as ampLayoutClassic} from '#builtins/amp-layout/build-dom';
 
-import {buildDom as ampFitTextClassic} from '../../extensions/amp-fit-text/0.1/amp-fit-text';
+import {buildDom as ampFitTextClassic} from '../../extensions/amp-fit-text/0.1/build-dom';
 
 const builderMap = {
   'v0': {
