+++ src/compiler/index.js
@@ -1,3 +1,4 @@
+import './polyfills';
 import * as compiler from '@ampproject/bento-compiler';
 
 import {getBuilders} from './builders';
